+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Changelog for NeoFS Node
 - IR logs INFO its state on start based on read FS chain's internals (#3197)
 - IR cache size for handled notary requests (#3205)
 - Metabase V3->4 migration routine no longer breaks on broken binary or proto-violating object encounter (#3203)
+- Metabase V3->4 migration routine skips objects from the removed containers (#3203)
 
 ### Removed